llm_earnings_insights/requirements.txt
streamlit
pydantic
PyMuPDF
faiss-cpu
sentence-transformers
openai
ollama
anthropic
requests
